// frontend/src/lib/api-client/models/UserInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UserInput
 */
export interface UserInput {
    /**
     * 
     * @type {string}
     * @memberof UserInput
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof UserInput
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof UserInput
     */
    lastName?: string;
    /**
     * 
     * @type {string}
     * @memberof UserInput
     */
    username: string;
    /**
     * 
     * @type {string}
     * @memberof UserInput
     */
    password: string;
    /**
     * 
     * @type {string}
     * @memberof UserInput
     */
    email?: string;
    /**
     * 
     * @type {Date}
     * @memberof UserInput
     */
    dateOfBirth?: Date;
}

/**
 * Check if a given object implements the UserInput interface.
 */
export function instanceOfUserInput(value: object): value is UserInput {
    if (!('firstName' in value) || value['firstName'] === undefined) return false;
    if (!('username' in value) || value['username'] === undefined) return false;
    if (!('password' in value) || value['password'] === undefined) return false;
    return true;
}

export function UserInputFromJSON(json: any): UserInput {
    return UserInputFromJSONTyped(json, false);
}

export function UserInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserInput {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'firstName': json['firstName'],
        'lastName': json['lastName'] == null ? undefined : json['lastName'],
        'username': json['username'],
        'password': json['password'],
        'email': json['email'] == null ? undefined : json['email'],
        'dateOfBirth': json['dateOfBirth'] == null ? undefined : (new Date(json['dateOfBirth'])),
    };
}

export function UserInputToJSON(json: any): UserInput {
    return UserInputToJSONTyped(json, false);
}

export function UserInputToJSONTyped(value?: UserInput | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'firstName': value['firstName'],
        'lastName': value['lastName'],
        'username': value['username'],
        'password': value['password'],
        'email': value['email'],
        'dateOfBirth': value['dateOfBirth'] == null ? undefined : ((value['dateOfBirth']).toISOString().substring(0,10)),
    };
}